Add files via upload
I am  not sure if we  should create another class for distributing the cards to the players , will figure it out today though.  except for the above so far only created a window for the cards( graphics) but i do not know how to combine everything together .  hmmm... hopefully i figure it out or u guys help.